# swagger/RedShirt.Example.Api.Interop/swagger.json
{
...
  "openapi": "3.0.0",
  "info": {
    "title": "RedShirt.Example.Api",
    "version": "1.0.0"
  },
  "paths": {
...
    "/messages/event-stream": {
      "get": {
        "tags": [
          "Message"
        ],
        "summary": "Event stream",
        "description": "This GET /messages/event-stream endpoint is not directly supported by OpenAPI.\nThe generated NSwag MessageClient.GetEventStreamAsync method is intended for connectivity checks only.\nUse MessagesEventStreamListener in RedShirt.Example.Api.Interop project to receive message events.",
        "operationId": "Message_GetEventStream",
        "responses": {
          "200": {
            "description": ""
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "502": {
            "description": ""
          }
        }
      }
    },
    "/messages": {
      "post": {
        "tags": [
          "Message"
        ],
        "operationId": "Message_Post",
        "requestBody": {
          "x-name": "request",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExampleMessagePostRequest"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "202": {
            "description": ""
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "502": {
            "description": ""
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ProblemDetails": {
        "type": "object",
        "additionalProperties": {
          "nullable": true
        },
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        }
      },
...
      "ExampleMessagePostRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "message"
        ],
        "properties": {
          "message": {
            "type": "string"
          }
        }
      },
...
    }
  }
}